requirements.txt
certifi==2025.4.26
contourpy==1.3.2
cycler==0.12.1
fitdecode==0.10.0
fonttools==4.58.1
geopandas==1.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
shapely==2.1.1
six==1.17.0
tzdata==2025.2
